requirements.txt
Django==2.0.7
django-phonenumber-field==2.0.0
